+++ packages/google_sign_in/google_sign_in_ios/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.6.4
+
+* Converts platform communication to Pigeon.
+
 ## 5.6.3
 
 * Adds pub topics to package metadata.